requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-dev.in -o requirements-dev.txt --universal
build==1.4.0
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via pip-tools
colorama==0.4.6 ; os_name == 'nt' or sys_platform == 'win32'
    # via
    #   build
    #   click
    #   pytest
coverage==7.10.7
    # via pytest-cov
debugpy==1.8.19
    # via -r requirements-dev.in
distlib==0.4.0
    # via virtualenv
exceptiongroup==1.3.1 ; python_full_version < '3.11'
    # via pytest
filelock==3.19.1 ; python_full_version < '3.10'
    # via virtualenv
filelock==3.20.3 ; python_full_version >= '3.10'
    # via virtualenv
flake8==7.3.0
    # via
    #   -r requirements-dev.in
    #   flake8-qgis
flake8-qgis==1.1.0
    # via -r requirements-dev.in
identify==2.6.15
    # via pre-commit
importlib-metadata==8.7.1 ; python_full_version < '3.10.2'
    # via build
iniconfig==2.1.0
    # via pytest
librt==0.7.8 ; platform_python_implementation != 'PyPy'
    # via mypy
mccabe==0.7.0
    # via flake8
mypy==1.19.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via mypy
nodeenv==1.10.0
    # via pre-commit
packaging==26.0
    # via
    #   build
    #   pytest
    #   wheel
pathspec==1.0.3
    # via mypy
pip==25.3
    # via pip-tools
pip-tools==7.5.2
    # via -r requirements-dev.in
platformdirs==4.4.0
    # via virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.3.0
    # via -r requirements-dev.in
pycodestyle==2.14.0
    # via flake8
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyqt5-stubs==5.15.6.0
    # via
    #   -r requirements-dev.in
    #   qgis-stubs
pytest==8.4.2
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-qgis
pytest-cov==7.0.0
    # via -r requirements-dev.in
pytest-qgis==2.1.0
    # via -r requirements-dev.in
pyyaml==6.0.3
    # via pre-commit
qgis-plugin-tools==0.5.0
    # via -r requirements-dev.in
qgis-stubs==0.2.0.post1
    # via -r requirements-dev.in
ruff==0.14.13
    # via -r requirements-dev.in
setuptools==80.10.1
    # via pip-tools
tomli==2.4.0 ; python_full_version <= '3.11'
    # via
    #   build
    #   coverage
    #   mypy
    #   pip-tools
    #   pytest
typing-extensions==4.15.0
    # via
    #   exceptiongroup
    #   mypy
    #   virtualenv
virtualenv==20.36.1
    # via pre-commit
wheel==0.46.2
    # via pip-tools
zipp==3.23.0 ; python_full_version < '3.10.2'
    # via importlib-metadata
